==> codes/01_Importing_Packages.Rmd <==
---
title: "Importing Packages"
output: github_document
---

Loading Packages
----------------
```{r Installing Packages, message=FALSE, warning=FALSE}
# Suppressing start-up messages and warnings for cleaner output.
suppressPackageStartupMessages({
  
  # Load the tidyverse package, which includes ggplot2, dplyr, tidyr, lubridate, stringr, readr.
  library(tidyverse)
  
  # Load the janitor package for data cleaning functions.
  library(janitor)
  
  # Load the scales package for data scaling and formatting.
  library(scales)
  
  # Load the knitr package for dynamic report generation.
  library(knitr)
  
  # Load the here package for managing file paths.
  library(here)
  
  # Load the ggdark package for creating dark-themed plots with ggplot2.
  library(ggdark)
})
```
